feat: 다른 사용자의 변경된 정보 가져옴
