+++ frontend/app/src/main/java/kr/ac/tukorea/whereareu/di/RepositoryModule.kt
@@ -41,7 +41,7 @@ object RepositoryModule {
     fun providesSettingRepository(
         modifyUserInfoService: ModifyUserInfoService
     ): SettingRepository = SettingRepositoryImpl(modifyUserInfoService)
-    
+
     @ViewModelScoped
     @Provides
     fun providesUserInfoRepository(
