+++ frontend/app/src/main/java/kr/ac/tukorea/whereareu/presentation/nok/setting/NokSettingFragment.kt
@@ -7,6 +7,7 @@ import androidx.navigation.fragment.findNavController
 import dagger.hilt.android.AndroidEntryPoint
 import kr.ac.tukorea.whereareu.R
 import kr.ac.tukorea.whereareu.databinding.FragmentNokSettingBinding
+import kr.ac.tukorea.whereareu.presentation.UserInfoViewModel
 import kr.ac.tukorea.whereareu.presentation.base.BaseFragment
 import kr.ac.tukorea.whereareu.presentation.nok.home.NokHomeViewModel
 import kr.ac.tukorea.whereareu.util.extension.repeatOnStarted
@@ -14,7 +15,7 @@ import kr.ac.tukorea.whereareu.util.extension.repeatOnStarted
 @AndroidEntryPoint
 class NokSettingFragment: BaseFragment<FragmentNokSettingBinding>(R.layout.fragment_nok_setting) {
     private val viewModel: NokHomeViewModel by activityViewModels()
-    private val settingViewModel: SettingViewModel by activityViewModels()
+    private val userInfoViewModel: UserInfoViewModel by activityViewModels()
 
     override fun initObserver() {
     }
@@ -53,7 +54,7 @@ class NokSettingFragment: BaseFragment<FragmentNokSettingBinding>(R.layout.fragm
         Log.d("settingFragment", "onResume")
 
         repeatOnStarted {
-            settingViewModel.userInfo.collect{
+            userInfoViewModel.userInfo.collect{
                 Log.d("Nok_Setting_Fragment", "get User Info API")
                 val nokName = it.result.nokInfoRecord.nokName
                 val nokPhone = it.result.nokInfoRecord.nokPhoneNumber
